refactor(amm): extract admin endpoints into dedicated module
Pure code move from src/rumi_amm/src/lib.rs to src/rumi_amm/src/admin.rs.
No behavior change, no signature change, no Candid change. Reduces
lib.rs by ~487 lines by extracting the 12 admin-gated endpoints into
their own module.

Follow-up to PR #193 code-quality review.
